(#13) Improve how provider handles unset config settings
Prior to this commit, this plugin would blow up if certain config
settings were not properly set by the user within their Vagrantfile or
in the vmfloaty config file. This commit adds some additional checks to
the config parsing as well as making unset config settings to nil
instead of a ruby Object.